Add rust-sensors app
This is a no frills Rust app that mimics the periodic update and observation
of sensor data. The goal is to focus the app on the basic usage of the Ditto
SDK without all the noise usually introduced by integrating the Distto
SDK with external frameworks. This would allow new users to focus directly on
how to use the Ditto SDK.

Run multiple instances of the app in different terminals, e.g.:
$ cargo run -- --id kitchen
$ cargo run -- --id bedroom"