fix: use correct Flutter version format in update-sdk-versions workflow
The workflow was using flutter-version: "stable" which caused the action
to interpret "stable" as both the version and channel, resulting in the
error "Unable to determine Flutter version for channel: stable version: stable".

Fixed by using flutter-version: '3.x' with channel: 'stable', matching
the pattern used in the flutter-ci.yml workflow.